5489: Bulk Download: Add generic odp get datum
