+++ .run/db_backup-import.run.xml
@@ -1,7 +1,8 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="db:backup-import" type="ShConfigurationType">
-    <option name="SCRIPT_TEXT" value="(cd src/tools/db/importBackup &amp;&amp; ./exec.sh) &amp;&amp; ts-node src/tools/generateCache/index.ts" />
+    <option name="SCRIPT_TEXT" value="(cd src/tools/db/importBackup &amp;&amp; ./exec.sh) &amp;&amp; ts-node src/tools/generateCache/index.ts &amp;&amp; ts-node src/tools/createMeasuresDimensions/index.ts" />
     <option name="INDEPENDENT_SCRIPT_PATH" value="true" />
+    <option name="SCRIPT_PATH" value="" />
     <option name="SCRIPT_OPTIONS" value="" />
     <option name="INDEPENDENT_SCRIPT_WORKING_DIRECTORY" value="true" />
     <option name="SCRIPT_WORKING_DIRECTORY" value="$PROJECT_DIR$" />
