+++ src/server/controller/cycleData/getBulkDownload/metadata/_files/ndpYear.ts
@@ -1,13 +1,10 @@
 import { ODPDataSourceMethod } from 'meta/assessment/originalDataPoint'
-import { Objects } from 'utils/objects'
 
 import { BulkDownloadFileFactory } from 'server/controller/cycleData/getBulkDownload/metadata/_types'
 import {
   BulkDownloadColNode,
   BulkDownloadColType,
   BulkDownloadDatumType,
-  BulkDownloadGetDatum,
-  BulkDownloadODPCountryData,
   BulkDownloadRow,
 } from 'server/controller/cycleData/getBulkDownload/types'
 
@@ -18,31 +15,11 @@ const colType = BulkDownloadColType.odp
 export const getNDPYear: BulkDownloadFileFactory = (props) => {
   const { i18n } = props
 
-  const getDatum: BulkDownloadGetDatum = (props) => {
-    const { colName, countryIso, data, variableName } = props
-    const countryData = data.odp[countryIso]
-
-    if (Objects.isNil(countryData)) return null
-
-    const tableKey = variableName as keyof BulkDownloadODPCountryData
-    const value = countryData[tableKey]
-
-    if (tableKey === 'dataSourceMethods') {
-      const method = colName as ODPDataSourceMethod
-      const included = Boolean(value?.includes(method))
-      return i18n.t(`yesNoTextSelect.${included ? 'yes' : 'no'}`)
-    }
-
-    return value as string
-  }
-
   const colNodes: Array<BulkDownloadColNode> = [
     {
       colName: 'minYear',
       colType,
       csvColumn: i18n.t('bulkDownload.NDPYear.earliestYear'),
-      datumType,
-      getDatum,
       tableName,
       variableName: 'minYear',
     },
@@ -51,7 +28,6 @@ export const getNDPYear: BulkDownloadFileFactory = (props) => {
       colType,
       csvColumn: i18n.t('bulkDownload.NDPYear.latestYear'),
       datumType,
-      getDatum,
       tableName,
       variableName: 'maxYear',
     },
@@ -64,7 +40,6 @@ export const getNDPYear: BulkDownloadFileFactory = (props) => {
             ? i18n.t('common.other')
             : i18n.t(`nationalDataPoint.dataSourceMethodsOptions.${colName}`),
         datumType,
-        getDatum,
         tableName,
         variableName: 'dataSourceMethods',
       }
@@ -73,5 +48,5 @@ export const getNDPYear: BulkDownloadFileFactory = (props) => {
 
   const row: BulkDownloadRow = { colNodes }
 
-  return { getDatum, includeClimaticDomain: true, includeForestArea: true, fileName: 'NDPYear', rows: [row] }
+  return { includeClimaticDomain: true, includeForestArea: true, fileName: 'NDPYear', rows: [row] }
 }
